Conflicts:
	PowerUpCore/PowerUp/Combos/StandardBuildTasks.ps1
	PowerUpCore/PowerUp/RunPSake.ps1

+++ PowerUpCore/PowerUp/Modules/PowerUpFileSystem/PowerUpFileSystem.psm1
@@ -149,6 +149,11 @@ function Test-ObjectFullName(
 function Remove-DirectoryFailSafe(
     [Parameter(Mandatory=$true)] [string] $path
 ) {
+    if (!(Test-Path $path)) {
+        Write-Host "Directory '$path' does not exist (no need to delete)."
+        return
+    }
+
     Remove-DirectoryFailSafeInternal (Get-Item $path)
 }
